requirements.txt
fastapi
uvicorn[standard]
feedparser
anthropic
apscheduler
python-dotenv
requests
aiofiles
